+++ src/me/itsatacoshop247/TreeAssist/blocklists/EmptyBlockList.java
@@ -1,9 +1,15 @@
 package me.itsatacoshop247.TreeAssist.blocklists;
 
+import me.itsatacoshop247.TreeAssist.core.Utils;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 
 public class EmptyBlockList implements BlockList {
+	protected int lookupTime;
+
+	public EmptyBlockList() {
+		lookupTime = Utils.plugin.getConfig().getInt("Placed Blocks.Handler Lookup Time", 60*60*24);
+	}
 
 	@Override
 	public void initiate() {}